Update5.0
This may be final version. It fix the problem that can not download
content of special stylesheet blogs. It adds the discrimination for
homepage or post page, which means that user can download whole blog or
specific post.

The main function is working for  lots of blogs, like special url or
theme. Of course, there may be some freak blogs’ stylesheet that is
incompatible. You are welcome to remind me if you have some find. :)